Add files via upload
Code for dataset generation. It refers to dns2tcpssh.txt and dns2tcpp2p.txt where the two applications are tunneled inside regular DNS. The code creates a mix of DNS (label 0 in the datasets) with tunnels (for ssh and p2p separately, label 1 in the datasets). Just replace dns2tcpp2p.txt with dns2tcpssh.txt in the main code to swhitch to the desired application. The "contatoreMix" variable in the code should be set as follows in order to obtain different mixtures of hidden application inside regular DNS: 1 corresponding to 50% - 9 10% - 99 1% - 999 0.1% - 5 20%.